requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
flask-wtf==1.2.1
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
Werkzeug==3.0.1
bcrypt==4.1.2
ollama>=0.1.0
groq>=0.11.0
flask-socketio>=5.3.6
eventlet>=0.33.3
python-socketio>=5.9.0
greenlet>=3.0.0
python-dotenv>=1.0.0
